feat: Add CurrentToast component for displaying toast notifications
The code changes add a new component called `CurrentToast` that is responsible for rendering toast notifications. It uses the `Toast` component from the `@tamagui/toast` library and displays the title and message of the current toast. The component also includes a `ToastControl` component that allows users to show and hide toast notifications using buttons.

This commit message follows the established convention of using a prefix to indicate the type of change (`feat` for feature). The message provides a clear and concise description of the code changes and their purpose.

Signed-off-by: David Hooton <david@hooton.org>