feat: Add ModalScreen component for displaying a modal
The code changes add a new component called `ModalScreen` that is responsible for rendering a modal. It uses the `tamagui` library to create a modal with a centered view and a stack of elements. The modal includes a paragraph and two anchor links. This component is intended to be used for displaying information or actions in a modal format.

Signed-off-by: David Hooton <david@hooton.org>