feat: Add TabLayout component for managing tabs
The code changes add a new component called `TabLayout` that is responsible for managing tabs using the `Tabs` component from the `expo-router` library. It includes two screens, "Tab One" and "Tab Two", with corresponding icons. The first screen also includes a button that navigates to a modal screen. This component enhances the app's navigation and provides a user-friendly tabbed interface.

Signed-off-by: David Hooton <david@hooton.org>