chore: Add .gitignore file
This commit adds a .gitignore file to the repository. The .gitignore file specifies which files and directories should be ignored by Git. It helps to prevent unnecessary files from being tracked and committed to the repository. The added .gitignore file includes patterns to ignore various files and directories such as node_modules, .next, .expo, build, dist, and more.

Signed-off-by: David Hooton <david@hooton.org>